feat: 404 page for wrong routes
