+++ src/__tests__/Search.test.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, RouterProvider } from 'react-router-dom';
 import Search from '../components/Search';
-import { SearchContext } from '../App';
+import { SearchContext } from '../pages/main';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { vi } from 'vitest';
