+++ src/router/index.tsx
@@ -3,13 +3,17 @@ import {
   createRoutesFromElements,
   Route,
 } from 'react-router-dom';
-import App from '../App';
 import Details from '../components/Details';
+import Error404 from '../pages/Error404';
+import App from '../App';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
-    <Route path="/" element={<App />}>
-      <Route index element={<Details />} />
+    <Route>
+      <Route path="/" element={<App />}>
+        <Route index element={<Details />} />
+      </Route>
+      <Route path="*" element={<Error404 />} />
     </Route>
   )
 );
